requirements.txt
attrs==22.2.0
Automat==22.10.0
bcrypt==4.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
constantly==15.1.0
defusedxml==0.7.1
exceptiongroup==1.1.1
Flask==2.0.2
Flask-Cors==3.0.10
Flask-Script==2.0.5
Flask-Twisted==0.1.2
greenlet==2.0.2
httpie==3.2.1
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.1.0
incremental==22.10.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.0.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
multidict==6.0.4
mysql-connector-python==8.0.32
observable==1.0.3
packaging==23.0
pluggy==1.0.0
protobuf==3.20.3
Pygments==2.14.0
PyJWT==2.6.0
PySocks==1.7.1
pytest==7.2.2
requests==2.28.2
requests-toolbelt==0.10.1
rich==13.3.3
six==1.16.0
SQLAlchemy==1.4.47
tomli==2.0.1
Twisted==22.10.0
typing-extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.0.3
zipp==3.15.0
zope.interface==6.0